requirements.txt
# Core dependencies
impacket>=0.9.24
ldap3>=2.9
colorama>=0.4.6

# Additional dependencies
dnspython>=2.4.2
pycryptodome>=3.20.0

# Optional (for advanced features)
rich>=13.7.0
cryptography>=42.0.0
